Comments can only be deleted by authentificated user.

+++ App/spring-backend/demo/src/main/java/com/example/demo/service/CommentService.java
@@ -3,6 +3,7 @@ package com.example.demo.service;
 import com.example.demo.dto.comment.CommentCreateDto;
 import com.example.demo.dto.comment.CommentUpdateDto;
 import com.example.demo.dto.user.UserUpdateDto;
+import com.example.demo.exception.AccessDeniedException;
 import com.example.demo.exception.CommentNotFoundException;
 import com.example.demo.exception.PostNotFoundException;
 import com.example.demo.exception.UserNotFoundException;
@@ -27,13 +28,12 @@ public class CommentService {
     private final CommentRepository commentRepository;
     private final UserRepository userRepository;
     private final PostRepository postRepository;
+    private final UserService userService;
 
     @Transactional
     public Comment addComment (CommentCreateDto commentDto) {
 
-        User authorUser = userRepository.findByUsername(commentDto.getAuthorUsername()).orElseThrow(() ->
-                new UserNotFoundException("User with username: " + commentDto.getAuthorUsername() +
-                        " was not found."));
+        User authorUser = userService.getAuthenticatedUser();
 
         Post targetPost = postRepository.findById(commentDto.getPostId()).orElseThrow(() ->
                 new PostNotFoundException("Post with id: " + commentDto.getPostId() +
@@ -62,6 +62,10 @@ public class CommentService {
     public void deleteCommentById (Long id) {
         Comment comment = commentRepository.findById(id).orElseThrow(
                 () -> new CommentNotFoundException("Comment with id: " + " was not found."));
+
+        if (!comment.getUser().equals(userService.getAuthenticatedUser()))
+            throw new AccessDeniedException("You are not the author of this comment");
+
         commentRepository.delete(comment);
     }
